requirements.txt
asgiref==3.5.2
Django==2.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.1
PyJWT==2.5.0
pytz==2022.4
django-random-id-model==0.1.1
django-cors-headers==2.2.0

sqlparse==0.4.3
